Fixed inconsistant use of filters by implementing generator filter variable.
Variable "self.df" was being correctly used to specify the number of filters in discriminator's layers. 
But the corresponding variable for the generator's layers "self.gf" was unimplemented and thus was an unused variable.

In this update I simply changed the function residual_block() to use the variable self.gf to correctly specify the number of filters in the blocks as intended.